renames files with `space` in the name
