+++ packages/insomnia-app/app/models/index.ts
@@ -21,6 +21,7 @@ import * as _grpcRequest from './grpc-request';
 import * as _grpcRequestMeta from './grpc-request-meta';
 import * as _oAuth2Token from './o-auth-2-token';
 import * as _pluginData from './plugin-data';
+import * as _space from './project';
 import * as _protoDirectory from './proto-directory';
 import * as _protoFile from './proto-file';
 import * as _request from './request';
@@ -30,7 +31,6 @@ import * as _requestMeta from './request-meta';
 import * as _requestVersion from './request-version';
 import * as _response from './response';
 import * as _settings from './settings';
-import * as _space from './space';
 import * as _stats from './stats';
 import * as _unitTest from './unit-test';
 import * as _unitTestResult from './unit-test-result';
