+++ packages/insomnia-app/app/sync/git/__tests__/ne-db-client.test.ts
@@ -6,7 +6,7 @@ import { globalBeforeEach } from '../../../__jest__/before-each';
 import { database as db } from '../../../common/database';
 import * as models from '../../../models';
 import { workspaceModelSchema } from '../../../models/__schemas__/model-schemas';
-import { BASE_PROJECT_ID } from '../../../models/space';
+import { BASE_PROJECT_ID } from '../../../models/project';
 import { GIT_CLONE_DIR, GIT_INSOMNIA_DIR, GIT_INSOMNIA_DIR_NAME } from '../git-vcs';
 import { NeDBClient } from '../ne-db-client';
 import { assertAsyncError, setupDateMocks } from './util';
