+++ packages/insomnia-app/app/sync/vcs/__tests__/initialize-project-from.test.ts
@@ -2,7 +2,7 @@ import { createBuilder } from '@develohpanda/fluent-builder';
 
 import { globalBeforeEach } from '../../../__jest__/before-each';
 import * as models from '../../../models';
-import { isRemoteSpace, Space } from '../../../models/space';
+import { isRemoteSpace, Space } from '../../../models/project';
 import { isCollection, Workspace, WorkspaceScopeKeys } from '../../../models/workspace';
 import { projectSchema, teamSchema } from '../../__schemas__/type-schemas';
 import { initializeSpaceFromTeam, initializeWorkspaceFromProject } from '../initialize-model-from';
