+++ packages/insomnia-app/app/sync/vcs/__tests__/pull-project.test.ts
@@ -4,7 +4,7 @@ import { mocked } from 'ts-jest/utils';
 import { globalBeforeEach } from '../../../__jest__/before-each';
 import { DEFAULT_BRANCH_NAME } from '../../../common/constants';
 import * as models from '../../../models';
-import { isRemoteSpace } from '../../../models/space';
+import { isRemoteSpace } from '../../../models/project';
 import { Workspace } from '../../../models/workspace';
 import { projectWithTeamSchema } from '../../__schemas__/type-schemas';
 import MemoryDriver from '../../store/drivers/memory-driver';
