+++ packages/insomnia-app/app/ui/components/dropdowns/workspace-dropdown.tsx
@@ -11,9 +11,9 @@ import { executeHotKey } from '../../../common/hotkeys-listener';
 import { RENDER_PURPOSE_NO_RENDER } from '../../../common/render';
 import { ApiSpec } from '../../../models/api-spec';
 import type { Environment } from '../../../models/environment';
+import { Space } from '../../../models/project';
 import { isRequest } from '../../../models/request';
 import { isRequestGroup } from '../../../models/request-group';
-import { Space } from '../../../models/space';
 import { isDesign, Workspace } from '../../../models/workspace';
 import type { WorkspaceAction } from '../../../plugins';
 import { ConfigGenerator, getConfigGenerators, getWorkspaceActions } from '../../../plugins';
