+++ packages/insomnia-app/app/ui/components/modals/project-settings-modal.tsx
@@ -6,9 +6,9 @@ import { bindActionCreators } from 'redux';
 import { AUTOBIND_CFG } from '../../../common/constants';
 import { strings } from '../../../common/strings';
 import * as models from '../../../models/index';
-import { isRemoteSpace, spaceHasSettings } from '../../../models/space';
+import { isRemoteSpace, spaceHasSettings } from '../../../models/project';
 import { RootState } from '../../redux/modules';
-import * as spaceActions from '../../redux/modules/space';
+import * as spaceActions from '../../redux/modules/project';
 import { selectActiveSpace } from '../../redux/selectors';
 import DebouncedInput from '../base/debounced-input';
 import Modal from '../base/modal';
