+++ packages/insomnia-app/app/ui/components/modals/sync-branches-modal.tsx
@@ -4,7 +4,7 @@ import React, { PureComponent } from 'react';
 
 import { AUTOBIND_CFG } from '../../../common/constants';
 import { database as db } from '../../../common/database';
-import { Space } from '../../../models/space';
+import { Space } from '../../../models/project';
 import type { Workspace } from '../../../models/workspace';
 import type { StatusCandidate } from '../../../sync/types';
 import { interceptAccessError } from '../../../sync/vcs/util';
