+++ packages/insomnia-app/app/ui/components/wrapper-home.tsx
@@ -26,7 +26,7 @@ import { isNotNullOrUndefined } from '../../common/misc';
 import { descendingNumberSort, sortMethodMap } from '../../common/sorting';
 import { strings } from '../../common/strings';
 import { ApiSpec } from '../../models/api-spec';
-import { isRemoteSpace } from '../../models/space';
+import { isRemoteSpace } from '../../models/project';
 import { isDesign, Workspace, WorkspaceScopeKeys } from '../../models/workspace';
 import { WorkspaceMeta } from '../../models/workspace-meta';
 import { MemClient } from '../../sync/git/mem-client';
@@ -40,9 +40,9 @@ import { activateWorkspace, createWorkspace } from '../redux/modules/workspace';
 import { selectSpaceSortOrder } from '../redux/selectors';
 import SettingsButton from './buttons/settings-button';
 import AccountDropdown from './dropdowns/account-dropdown';
+import { SpaceDropdown } from './dropdowns/project-dropdown';
+import { SpaceSortDropdown } from './dropdowns/project-sort-dropdown';
 import { RemoteWorkspacesDropdown } from './dropdowns/remote-workspaces-dropdown';
-import { SpaceDropdown } from './dropdowns/space-dropdown';
-import { SpaceSortDropdown } from './dropdowns/space-sort-dropdown';
 import KeydownBinder from './keydown-binder';
 import { showPrompt } from './modals';
 import Notice from './notice';
