+++ packages/insomnia-app/app/ui/redux/modules/__tests__/project.test.js
@@ -6,10 +6,10 @@ import { reduxStateForTest } from '../../../../__jest__/redux-state-for-test';
 import { SegmentEvent, trackEvent, trackSegmentEvent } from '../../../../common/analytics';
 import { ACTIVITY_HOME } from '../../../../common/constants';
 import * as models from '../../../../models';
-import { BASE_PROJECT_ID } from '../../../../models/space';
+import { BASE_PROJECT_ID } from '../../../../models/project';
 import { getAndClearShowAlertMockArgs, getAndClearShowPromptMockArgs } from '../../../../test-utils';
 import { SET_ACTIVE_ACTIVITY, SET_ACTIVE_SPACE } from '../global';
-import { createSpace, removeSpace } from '../space';
+import { createSpace, removeSpace } from '../project';
 
 jest.mock('../../../components/modals');
 jest.mock('../../../../common/analytics');
