+++ packages/insomnia-app/app/ui/redux/modules/entities.ts
@@ -13,6 +13,7 @@ import { GrpcRequest } from '../../../models/grpc-request';
 import { GrpcRequestMeta } from '../../../models/grpc-request-meta';
 import { OAuth2Token } from '../../../models/o-auth-2-token';
 import { PluginData } from '../../../models/plugin-data';
+import { Space } from '../../../models/project';
 import { ProtoDirectory } from '../../../models/proto-directory';
 import { ProtoFile } from '../../../models/proto-file';
 import { Request } from '../../../models/request';
@@ -22,7 +23,6 @@ import { RequestMeta } from '../../../models/request-meta';
 import { RequestVersion } from '../../../models/request-version';
 import { Response } from '../../../models/response';
 import { Settings } from '../../../models/settings';
-import { Space } from '../../../models/space';
 import { Stats } from '../../../models/stats';
 import { UnitTest } from '../../../models/unit-test';
 import { UnitTestResult } from '../../../models/unit-test-result';
