+++ packages/insomnia-app/app/ui/redux/modules/global.tsx
@@ -29,9 +29,9 @@ import * as models from '../../../models';
 import { Environment, isEnvironment } from '../../../models/environment';
 import { GrpcRequest } from '../../../models/grpc-request';
 import * as requestOperations from '../../../models/helpers/request-operations';
+import { BASE_PROJECT_ID } from '../../../models/project';
 import { Request } from '../../../models/request';
 import { Settings } from '../../../models/settings';
-import { BASE_PROJECT_ID } from '../../../models/space';
 import { isWorkspace } from '../../../models/workspace';
 import { reloadPlugins } from '../../../plugins';
 import { createPlugin } from '../../../plugins/create';
